float -> double

+++ Teensy/include/Motor/asservissement.hpp
@@ -6,23 +6,23 @@ class Asservissement{
         /**
          * @brief Construct a new Asservissement object
          */
-        Asservissement (float kp, float ki, float kd);
+        Asservissement (double kp, double ki, double kd);
         Asservissement () {};
 
         /**
          * @brief Computes the control signals for the robotic system based on the PID controllers and desired setpoints
          * @param dt Time elapsed since the last control signal was computed
          */
-        float compute (float error, unsigned long dt);
+        double compute (double error, unsigned long dt);
     
     private:
-        float kp;  // Proportional gain
-        float ki;  // Integral gain
-        float kd;  // Derivative gain
+        double kp;  // Proportional gain
+        double ki;  // Integral gain
+        double kd;  // Derivative gain
 
-        float previousError;
-        float integralError;
-        float derivativeError;
+        double previousError;
+        double integralError;
+        double derivativeError;
 
 };
